collision checks done!

+++ src/shapes/Circle.java
@@ -6,7 +6,7 @@ public class Circle extends Shape {
         randomProperties(bound);
     }
 
-    public Circle(int x, int y, float radius) {
+    public Circle(float x, float y, float radius) {
         this.x = x;
         this.y = y;
         this.radius = radius;
